Вынесла функцию parseDate в utils
